# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Scribe local transcription app API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Tag {
  id: number;
  name: string;
  color: string;
}

export interface Transcript {
  id: number;
  title: string;
  rawText: string;
  /** @nullable */
  cleanedText?: string | null;
  /** @nullable */
  summary?: string | null;
  /** @nullable */
  durationSeconds?: number | null;
  wordCount?: number;
  tags?: Tag[];
  createdAt: string;
  updatedAt: string;
}

export interface TranscriptInput {
  title?: string;
  rawText: string;
  cleanedText?: string;
  durationSeconds?: number;
  tagIds?: number[];
}

export interface TranscriptUpdate {
  title?: string;
  rawText?: string;
  cleanedText?: string;
  summary?: string;
  tagIds?: number[];
}

export type TranscriptStatsTopTagsItem = {
  name: string;
  count: number;
};

export interface TranscriptStats {
  totalCount: number;
  totalWords: number;
  totalDurationSeconds: number;
  topTags: TranscriptStatsTopTagsItem[];
  recentCount?: number;
}

export interface TagInput {
  name: string;
  color?: string;
}

export interface ExportResult {
  content: string;
  filename: string;
}

export interface Settings {
  /** @nullable */
  groqApiKey?: string | null;
  groqModel?: string;
  storageDir?: string;
}

export interface SettingsUpdate {
  groqApiKey?: string;
  groqModel?: string;
  storageDir?: string;
}

export type ListTranscriptsParams = {
search?: string;
tag?: string;
limit?: number;
offset?: number;
};