# lib/api-zod/src/generated/types/settings.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Scribe local transcription app API
 * OpenAPI spec version: 0.1.0
 */

export interface Settings {
  /** @nullable */
  groqApiKey?: string | null;
  groqModel?: string;
  storageDir?: string;
}