# lib/api-zod/src/generated/types/settingsUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Scribe local transcription app API
 * OpenAPI spec version: 0.1.0
 */

export interface SettingsUpdate {
  groqApiKey?: string;
  groqModel?: string;
  storageDir?: string;
}